new test script

+++ quad/gcd-coverage.c
@@ -7,12 +7,12 @@ void *_coverage_fout ;
 /* compiler builtin: 
    int __builtin_strcmp(char const   * , char const   * ) ;  */
 /* compiler builtin: 
-   void *__builtin___memmove_chk(void * , void const   * , unsigned int  ,
-                                 unsigned int  ) ;  */
+   void *__builtin___memmove_chk(void * , void const   * , unsigned long  ,
+                                 unsigned long  ) ;  */
 /* compiler builtin: 
    char *__builtin_strpbrk(char const   * , char const   * ) ;  */
 /* compiler builtin: 
-   void *__builtin_memcpy(void * , void const   * , unsigned int  ) ;  */
+   void *__builtin_memcpy(void * , void const   * , unsigned long  ) ;  */
 /* compiler builtin: 
    double __builtin_exp(double  ) ;  */
 /* compiler builtin: 
@@ -24,8 +24,8 @@ void *_coverage_fout ;
 /* compiler builtin: 
    float __builtin_atan2f(float  , float  ) ;  */
 /* compiler builtin: 
-   void *__builtin___memcpy_chk(void * , void const   * , unsigned int  ,
-                                unsigned int  ) ;  */
+   void *__builtin___memcpy_chk(void * , void const   * , unsigned long  ,
+                                unsigned long  ) ;  */
 /* compiler builtin: 
    double __builtin_asin(double  ) ;  */
 /* compiler builtin: 
@@ -41,21 +41,21 @@ void *_coverage_fout ;
 /* compiler builtin: 
    int __builtin___fprintf_chk(void * , int  , char const   *  , ...) ;  */
 /* compiler builtin: 
-   int __builtin___vsprintf_chk(char * , int  , unsigned int  , char const   * ,
-                                __builtin_va_list  ) ;  */
+   int __builtin___vsprintf_chk(char * , int  , unsigned long  ,
+                                char const   * , __builtin_va_list  ) ;  */
 /* compiler builtin: 
-   char *__builtin___strncpy_chk(char * , char const   * , unsigned int  ,
-                                 unsigned int  ) ;  */
+   char *__builtin___strncpy_chk(char * , char const   * , unsigned long  ,
+                                 unsigned long  ) ;  */
 /* compiler builtin: 
    float __builtin_log10f(float  ) ;  */
 /* compiler builtin: 
    double __builtin_atan(double  ) ;  */
 /* compiler builtin: 
-   void *__builtin_alloca(unsigned int  ) ;  */
+   void *__builtin_alloca(unsigned long  ) ;  */
 /* compiler builtin: 
    void __builtin_va_end(__builtin_va_list  ) ;  */
 /* compiler builtin: 
-   int __builtin_strncmp(char const   * , char const   * , unsigned int  ) ;  */
+   int __builtin_strncmp(char const   * , char const   * , unsigned long  ) ;  */
 /* compiler builtin: 
    double __builtin_sin(double  ) ;  */
 /* compiler builtin: 
@@ -63,10 +63,10 @@ void *_coverage_fout ;
 /* compiler builtin: 
    float __builtin_coshf(float  ) ;  */
 /* compiler builtin: 
-   void *__builtin___mempcpy_chk(void * , void const   * , unsigned int  ,
-                                 unsigned int  ) ;  */
+   void *__builtin___mempcpy_chk(void * , void const   * , unsigned long  ,
+                                 unsigned long  ) ;  */
 /* compiler builtin: 
-   char *__builtin___strcat_chk(char * , char const   * , unsigned int  ) ;  */
+   char *__builtin___strcat_chk(char * , char const   * , unsigned long  ) ;  */
 /* compiler builtin: 
    float __builtin_nansf(char const   * ) ;  */
 /* compiler builtin: 
@@ -118,9 +118,9 @@ void *_coverage_fout ;
 /* compiler builtin: 
    float __builtin_fmodf(float  ) ;  */
 /* compiler builtin: 
-   unsigned int __builtin_object_size(void * , int  ) ;  */
+   unsigned long __builtin_object_size(void * , int  ) ;  */
 /* compiler builtin: 
-   void __builtin_va_arg(__builtin_va_list  , unsigned int  , void * ) ;  */
+   void __builtin_va_arg(__builtin_va_list  , unsigned long  , void * ) ;  */
 /* compiler builtin: 
    void __builtin_stdarg_start(__builtin_va_list  ) ;  */
 /* compiler builtin: 
@@ -148,17 +148,17 @@ void *_coverage_fout ;
 /* compiler builtin: 
    int __builtin_popcountll(unsigned long long  ) ;  */
 /* compiler builtin: 
-   int __builtin___sprintf_chk(char * , int  , unsigned int  , char const   * 
+   int __builtin___sprintf_chk(char * , int  , unsigned long  , char const   * 
                                , ...) ;  */
 /* compiler builtin: 
    int __builtin___vprintf_chk(int  , char const   * , __builtin_va_list  ) ;  */
 /* compiler builtin: 
-   int __builtin___snprintf_chk(char * , unsigned int  , int  , unsigned int  ,
-                                char const   *  , ...) ;  */
+   int __builtin___snprintf_chk(char * , unsigned long  , int  ,
+                                unsigned long  , char const   *  , ...) ;  */
 /* compiler builtin: 
    long double __builtin_infl(void) ;  */
 /* compiler builtin: 
-   void *__builtin_mempcpy(void * , void const   * , unsigned int  ) ;  */
+   void *__builtin_mempcpy(void * , void const   * , unsigned long  ) ;  */
 /* compiler builtin: 
    long double __builtin_floorl(long double  ) ;  */
 /* compiler builtin: 
@@ -180,14 +180,15 @@ void *_coverage_fout ;
 /* compiler builtin: 
    float __builtin_acosf(float  ) ;  */
 /* compiler builtin: 
-   int __builtin___vsnprintf_chk(char * , unsigned int  , int  , unsigned int  ,
-                                 char const   * , __builtin_va_list  ) ;  */
+   int __builtin___vsnprintf_chk(char * , unsigned long  , int  ,
+                                 unsigned long  , char const   * ,
+                                 __builtin_va_list  ) ;  */
 /* compiler builtin: 
    double __builtin_sinh(double  ) ;  */
 /* compiler builtin: 
    int __builtin_ffsll(unsigned long long  ) ;  */
 /* compiler builtin: 
-   char *__builtin___strcpy_chk(char * , char const   * , unsigned int  ) ;  */
+   char *__builtin___strcpy_chk(char * , char const   * , unsigned long  ) ;  */
 /* compiler builtin: 
    double __builtin_inf(void) ;  */
 /* compiler builtin: 
@@ -207,7 +208,7 @@ void *_coverage_fout ;
 /* compiler builtin: 
    void *__builtin_return_address(unsigned int  ) ;  */
 /* compiler builtin: 
-   char *__builtin___stpcpy_chk(char * , char const   * , unsigned int  ) ;  */
+   char *__builtin___stpcpy_chk(char * , char const   * , unsigned long  ) ;  */
 /* compiler builtin: 
    float __builtin_tanf(float  ) ;  */
 /* compiler builtin: 
@@ -215,7 +216,7 @@ void *_coverage_fout ;
 /* compiler builtin: 
    float __builtin_ldexpf(float  , int  ) ;  */
 /* compiler builtin: 
-   int __builtin_types_compatible_p(unsigned int  , unsigned int  ) ;  */
+   int __builtin_types_compatible_p(unsigned long  , unsigned long  ) ;  */
 /* compiler builtin: 
    double __builtin_log10(double  ) ;  */
 /* compiler builtin: 
@@ -239,10 +240,10 @@ void *_coverage_fout ;
 /* compiler builtin: 
    int __builtin_popcountl(unsigned long  ) ;  */
 /* compiler builtin: 
-   void *__builtin___memset_chk(void * , int  , unsigned int  , unsigned int  ) ;  */
+   void *__builtin___memset_chk(void * , int  , unsigned long  , unsigned long  ) ;  */
 /* compiler builtin: 
-   char *__builtin___strncat_chk(char * , char const   * , unsigned int  ,
-                                 unsigned int  ) ;  */
+   char *__builtin___strncat_chk(char * , char const   * , unsigned long  ,
+                                 unsigned long  ) ;  */
 /* compiler builtin: 
    double __builtin_huge_val(void) ;  */
 /* compiler builtin: 
@@ -257,13 +258,13 @@ void *_coverage_fout ;
 /* compiler builtin: 
    double __builtin_atan2(double  , double  ) ;  */
 /* compiler builtin: 
-   char *__builtin_strncpy(char * , char const   * , unsigned int  ) ;  */
+   char *__builtin_strncpy(char * , char const   * , unsigned long  ) ;  */
 /* compiler builtin: 
    long double __builtin_powil(long double  , int  ) ;  */
 /* compiler builtin: 
    float __builtin_cosf(float  ) ;  */
 /* compiler builtin: 
-   unsigned int __builtin_strspn(char const   * , char const   * ) ;  */
+   unsigned long __builtin_strspn(char const   * , char const   * ) ;  */
 /* compiler builtin: 
    long double __builtin_modfl(long double  , long double * ) ;  */
 /* compiler builtin: 
@@ -271,7 +272,7 @@ void *_coverage_fout ;
 /* compiler builtin: 
    double __builtin_cosh(double  ) ;  */
 /* compiler builtin: 
-   char *__builtin_strncat(char * , char const   * , unsigned int  ) ;  */
+   char *__builtin_strncat(char * , char const   * , unsigned long  ) ;  */
 /* compiler builtin: 
    long __builtin_expect(long  , long  ) ;  */
 /* compiler builtin: 
@@ -283,7 +284,7 @@ void *_coverage_fout ;
 /* compiler builtin: 
    int __builtin_ctzll(unsigned long long  ) ;  */
 /* compiler builtin: 
-   unsigned int __builtin_strcspn(char const   * , char const   * ) ;  */
+   unsigned long __builtin_strcspn(char const   * , char const   * ) ;  */
 /* compiler builtin: 
    int __builtin_ffsl(unsigned long  ) ;  */
 /* compiler builtin:
